Move ctu/statistics capable functions under analyzer
Move `is_ctu_capable` and `is_statistics_capable` from common place
under the analyze module because these functions are only used in the
analyze part of the CodeChecker.
