+++ libcodechecker/analyze/analyzers/analyzer_types.py
@@ -10,7 +10,10 @@ Supported analyzer types.
 from __future__ import print_function
 from __future__ import division
 from __future__ import absolute_import
+
 import os
+import re
+import subprocess
 
 from libcodechecker.analyze import analyzer_env
 from libcodechecker.analyze import host_check
@@ -22,6 +25,36 @@ LOG = get_logger('analyzer')
 
 supported_analyzers = {ClangSA.ANALYZER_NAME: ClangSA,
                        ClangTidy.ANALYZER_NAME: ClangTidy}
+
+
+def is_ctu_capable(context):
+    """ Detects if the current clang is CTU compatible. """
+    ctu_func_map_cmd = context.ctu_func_map_cmd
+    try:
+        version = subprocess.check_output([ctu_func_map_cmd, '-version'])
+    except (subprocess.CalledProcessError, OSError):
+        version = 'ERROR'
+    return version != 'ERROR'
+
+
+def is_statistics_capable(context):
+    """ Detects if the current clang is Statistics compatible. """
+    # Resolve potentially missing binaries.
+    check_supported_analyzers([ClangSA.ANALYZER_NAME], context)
+    clangsa_cfg = ClangSA.construct_config_handler([], context)
+
+    check_env = analyzer_env.get_check_env(context.path_env_extra,
+                                           context.ld_lib_path_extra)
+
+    checkers = ClangSA.get_analyzer_checkers(clangsa_cfg, check_env)
+
+    stat_checkers_pattern = re.compile(r'.+statisticscollector.+')
+
+    for checker_name, _ in checkers:
+        if stat_checkers_pattern.match(checker_name):
+            return True
+
+    return False
 
 
 def check_supported_analyzers(analyzers, context):
